chore: delete old ComponentSettingsModal and rename new one
