+++ src/components/CollectionPagesSection.jsx
@@ -5,7 +5,7 @@ import _ from 'lodash';
 
 // Import components
 import OverviewCard from '../components/OverviewCard';
-import ComponentSettingsModal from '../components/NewComponentSettingsModal'
+import ComponentSettingsModal from './ComponentSettingsModal'
 
 // Import styles
 import elementStyles from '../styles/isomer-cms/Elements.module.scss';
