+++ src/components/ComponentSettingsModal.jsx
@@ -1,5 +1,6 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import axios from 'axios';
+import AsyncCreatableSelect from "react-select/async-creatable";
 import CreatableSelect from 'react-select/creatable'
 import { Base64 } from 'js-base64';
 import PropTypes from 'prop-types';
@@ -10,6 +11,7 @@ import {
   frontMatterParser,
   dequoteString,
   generatePageFileName,
+  generateCollectionPageFileName,
   generateResourceFileName,
   retrieveCollectionAndLinkFromPermalink,
   saveFileAndRetrieveUrl,
@@ -23,390 +25,440 @@ import SaveDeleteButtons from './SaveDeleteButtons';
 // axios settings
 axios.defaults.withCredentials = true
 
-export default class ComponentSettingsModal extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      title: '',
-      permalink: '',
-      fileUrl: null,
-      date: '',
-      mdBody: null,
-      category: '',
-      sha: '',
-      allCategories: null,
-      type: '',
-      errors: {
-        title: '',
-        permalink: '',
-        fileUrl: '',
-        date: '',
-        category: '',
-        originalCategory: '',
-      },
-      canShowDeleteWarningModal: false,
-      baseApiUrl: '',
-      originalCategory: '',
-      thirdNavTitle: '',
-      redirectToNewPage: false,
-      newPageUrl: '',
-      isPost: true,
-      originalPermalink: '',
-      originalFileUrl: '',
-    };
-  }
-
-  async componentDidMount() {
-    try {
-      const {
-        category, siteName, fileName, isNewFile, type
-      } = this.props;
-      
-      // Retrieve the list of all page/resource categories for use in the dropdown options. Also sets the default category if none is specified.
-      if (type === "resource") {
-        const resourcesResp = await axios.get(`${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/resources`);
-        const { resources: allCategories } = resourcesResp.data;
-        this.setState({
-          category: category ? category : '',
-          allCategories: allCategories.map((category) => category.dirName),
-        })
-      } else if (type === "page") {
-        const collectionsResp = await axios.get(`${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/collections`);
-        const { collections: collectionCategories } = collectionsResp.data;
-        const allCategories = [''].concat(collectionCategories)
-        this.setState({
-          category: category ? category : '',
-          allCategories,
-        })
-      }
-
-      const baseApiUrl = `${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}${category ? type === "resource" ? `/resources/${category}` : `/collections/${category}` : ''}`
-      if (isNewFile) {
-        // Set default values for a new file
-        this.setState({
-          title: 'Title',
-          permalink: 'permalink',
-          mdBody: '',
-          baseApiUrl,
-          type,
-          originalCategory: category,
-        });
-        // Only resources have a date field
-        if (type === "resource") this.setState({date:new Date().toISOString().split("T")[0]})
-      } else {
-        // Retrieve data from existing page/resource
-        const resp = await axios.get(`${baseApiUrl}/pages/${fileName}`);
-
-        const { content, sha } = resp.data;
-        const base64DecodedContent = Base64.decode(content);
-        const { frontMatter, mdBody } = frontMatterParser(base64DecodedContent);
-        const {
-          title, permalink, file_url: fileUrl, date, third_nav_title: thirdNavTitle,
-        } = frontMatter;
-
-        let existingLink
-        if (permalink) {
-          // We want to set parts of the link by default, and only leave a portion editable by the user
-          const { editableLink } = retrieveCollectionAndLinkFromPermalink(permalink)
-          existingLink = editableLink
-        }
-
-        this.setState({
-          title: dequoteString(title),
-          permalink: existingLink,
-          originalPermalink: existingLink,
-          fileUrl,
-          originalFileUrl: fileUrl,
-          date,
-          sha,
-          mdBody,
-          originalCategory: category,
-          baseApiUrl,
-          type,
-          thirdNavTitle,
-          isPost: !fileUrl
-        });
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  }
-
-  generateInitialCategoryLabel = (originalCategory, isCategoryDisabled) => {
+// Helper functions
+const generateInitialCategoryLabel = (originalCategory, isCategoryDisabled) => {
     if (originalCategory) return originalCategory
     // If category is disabled and no original category exists, it is an unlinked page
     return isCategoryDisabled ? "Unlinked Page" : "Select a category or create a new category..."
-  }
-
-  handlePermalinkFileUrlToggle = (event) => {
-    const { target: { value } } = event;
-    const { originalFileUrl, originalPermalink} = this.state
-    if (value === 'file') {
-      this.setState({
-        permalink: null,
-        fileUrl: originalFileUrl ? originalFileUrl : '',
-        isPost: false,
-        errors: {
-          permalink: '',
-        },
-      });
-    } else {
-      this.setState({
-        permalink: originalPermalink ? originalPermalink : 'permalink',
-        fileUrl: null,
-        isPost: true,
-        errors: {
-          fileUrl: '',
-        },
-      });
-    }
-  }
-
-  deleteHandler = async () => {
-    try {
-      const { sha, baseApiUrl } = this.state;
-      const { fileName } = this.props;
-      const params = { sha };
-      await axios.delete(`${baseApiUrl}/pages/${fileName}`, {
-        data: params,
-      });
-
-      // Refresh page
-      window.location.reload();
-    } catch (err) {
-      console.log(err);
-    }
-  }
-
-  saveHandler = async () => {
-    try {
-      const {
-        title, permalink, fileUrl, date, mdBody, sha, category, originalCategory, baseApiUrl, type, thirdNavTitle,
-      } = this.state;
-      const { fileName, isNewFile, siteName } = this.props;
-
-      const fileInfo = {
-        title,
-        permalink,
-        fileUrl,
-        date,
-        mdBody,
-        sha,
-        category,
-        originalCategory,
-        baseApiUrl,
-        type,
-        thirdNavTitle,
-        fileName,
-        isNewFile,
-        siteName,
-      }
-      
-      const newPageUrl = await saveFileAndRetrieveUrl(fileInfo)
-      // Refresh page
-      !isNewFile && window.location.reload();
-      this.setState({ redirectToNewPage: true, newPageUrl })
-    } catch (err) {
-      console.log(err);
-    }
-  }
-
-  changeHandler = (event) => {
-    const { id, value } = event.target;
-    const { title, date } = this.state
-    const { pageFilenames, type, fileName: currentFileName } = this.props;
-
-    const pageFilenamesExc = pageFilenames ? pageFilenames.filter((filename) => filename !== currentFileName) : null
-    let errorMessage, newFileName
-    if (type === 'resource') {
-      errorMessage = validateResourceSettings(id, value);
-      newFileName = generateResourceFileName(id==="title" ? value : title, id==="date" ? value : date)
-    } else if (type === 'page') {
-      errorMessage = validatePageSettings(id, value);
-      newFileName = generatePageFileName(value);
-    }
-
-    if (errorMessage === '' && pageFilenamesExc && pageFilenamesExc.includes(newFileName)) {
-      this.setState({
-        errors: {
-          title: 'This title is already in use. Please choose a different one.',
-        },
-        [id]: value,
-      })
-    } else {
-      this.setState({
-        errors: {
-          [id]: errorMessage,
-        },
-        [id]: value,
-      });
-    }
-  }
-
-  categoryDropdownHandler = (newValue) => {
-    let event
-    if (!newValue) {
-      // Field was cleared
-      event = {
-        target: {
-          id: 'category',
-          value: ''
-        }
-      }
-    } else {
-      const { value } = newValue
-      event = {
-        target: {
-          id: 'category',
-          value: value
-        }
-      }
-    }
-    
-    this.changeHandler(event)
-  };
-
-
-  render() {
-    const {
-      category,
-      title,
-      date,
-      allCategories,
-      permalink,
-      fileUrl,
-      errors,
-      canShowDeleteWarningModal,
-      sha,
-      redirectToNewPage,
-      newPageUrl,
-      thirdNavTitle,
-      isPost,
-    } = this.state;
-    const { settingsToggle, isNewFile, type, modalTitle, siteName, isCategoryDisabled, category:originalCategory } = this.props;
-
-    // Settings form has errors - disable save button
-    const hasErrors = _.some(errors, (field) => field.length > 0);
-
-    return (
-      <>
-        <div className={elementStyles.overlay}>
-          { (sha || isNewFile)
-          && (
-          <div className={elementStyles['modal-settings']}>
-            <div className={elementStyles.modalHeader}>
-              <h1>{modalTitle}</h1>
-              <button id="settings-CLOSE" type="button" onClick={settingsToggle}>
-                <i id="settingsIcon-CLOSE" className="bx bx-x" />
-              </button>
-            </div>
-            <div className={elementStyles.modalContent}>
-              <div className={elementStyles.modalFormFields}>
-                {/* Category */}
-                <p className={elementStyles.formLabel}>Category Folder Name</p>
-                <div className="d-flex text-nowrap">
-                  <CreatableSelect
-                    isClearable
-                    className="w-100"
-                    onChange={this.categoryDropdownHandler}
-                    isDisabled={isCategoryDisabled}
-                    defaultValue={{value:originalCategory,label:this.generateInitialCategoryLabel(originalCategory,isCategoryDisabled)}}
-                    options={
-                      allCategories
-                      ? allCategories.map((category) => (
-                        {
-                          value:category,
-                          label:category ? category : 'Unlinked Page'
-                        }
-                      ))
-                      : null
-                    }
-                  />
-                </div>
-                <span className={elementStyles.error}>{errors.category}</span>
-                {/* Title */}
-                <FormField
-                  title="Title"
-                  id="title"
-                  value={title}
-                  errorMessage={errors.title}
-                  isRequired={true}
-                  onFieldChange={this.changeHandler}
-                />
-                {/* Third Nav */}
-                { 
-                  thirdNavTitle &&
-                  <FormField
-                    title="3rd Nav Title"
-                    id="thirdNavTitle"
-                    value={thirdNavTitle}
-                    onFieldChange={this.changeHandler}
-                  />
-                }
-                {/* Permalink */}
-                <FormField
-                  title="Permalink"
-                  id="permalink"
-                  value={permalink ? permalink : ''}
-                  errorMessage={errors.permalink}
-                  isRequired={true}
-                  onFieldChange={this.changeHandler}
-                  disabled={!isPost}
-                  fixedMessage={category ? `/${category}/${thirdNavTitle ? `${thirdNavTitle}/` : ''}` : '/'}
-                  placeholder=' '
-                />
-                { type === "resource" && 
-                <ResourceFormFields 
-                  date={date}
-                  errors={errors}
-                  changeHandler={this.changeHandler}
-                  onToggle={this.handlePermalinkFileUrlToggle}
-                  isPost={isPost}
-                  siteName={siteName}
-                  fileUrl={fileUrl ? fileUrl : ''}
-                />
-                }
-              </div>
-              <SaveDeleteButtons 
-                isDisabled={isNewFile ? hasErrors || (type==='resource' && category==='') : (hasErrors || !sha)}
-                hasDeleteButton={!isNewFile}
-                saveCallback={this.saveHandler}
-                deleteCallback={() => this.setState({ canShowDeleteWarningModal: true })}
-              />
-            </div>
-          </div>
-          )}
-        </div>
-        {
-          canShowDeleteWarningModal
-          && (
-            <DeleteWarningModal
-              onCancel={() => this.setState({ canShowDeleteWarningModal: false })}
-              onDelete={this.deleteHandler}
-              type="resource"
-            />
-          )
-        }
-        {
-          isNewFile
-          && redirectToNewPage
-          && (
-            <Redirect
-              to={{
-                pathname: newPageUrl
-              }}
-            />
-          )
-        }
-      </>
-    );
-  }
 }
 
+const generateInitialThirdNavLabel = (thirdNavTitle, originalCategory) => {
+    if (thirdNavTitle) return thirdNavTitle
+    if (originalCategory && !thirdNavTitle) return 'None'
+    return "Select a third nav section..."
+}
+
+const ComponentSettingsModal = ({
+    category: originalCategory,
+    fileName,
+    isCategoryDisabled,
+    isNewFile,
+    modalTitle,
+    collectionPageData,
+    pageFileNames,
+    settingsToggle,
+    siteName,
+    type,
+    loadThirdNavOptions,
+}) => {
+    // Errors
+    const [errors, setErrors] = useState({
+        title: '',
+        permalink: '',
+        fileUrl: '',
+        resourceDate: '',
+        category: '',
+        originalCategory: '',
+    })
+    const [hasErrors, setHasErrors] = useState(false)
+
+    // Base hooks
+    const [allCategories, setAllCategories] = useState(null);
+    const [category, setCategory] = useState('');
+    const [baseApiUrl, setBaseApiUrl] = useState('');
+    const [title, setTitle] = useState('')
+    const [permalink, setPermalink] = useState('')
+    const [mdBody, setMdBody] = useState('')
+    const [sha, setSha] = useState('')
+    const [isPost, setIsPost] = useState(true)
+
+    // Track original values
+    const [originalPermalink, setOriginalPermalink] = useState('')
+    const [originalFileUrl, setOriginalFileUrl] = useState('')
+
+    // Collections-related
+    const [originalThirdNavTitle, setOriginalThirdNavTitle] = useState('') // this state is required for comparison purposes
+    const [thirdNavTitle, setThirdNavTitle] = useState('')
+
+    // Resource-related
+    const [resourceDate, setResourceDate] = useState('')
+    const [fileUrl, setFileUrl] = useState('')
+
+    // Page redirection and modals
+    const [newPageUrl, setNewPageUrl] = useState('')
+    const [redirectToNewPage, setRedirectToNewPage] = useState(false)
+    const [canShowDeleteWarningModal, setCanShowDeleteWarningModal] = useState(false)
+
+    // Map element ID to setter functions
+    const idToSetterFuncMap = {
+        category: setCategory,
+        title: setTitle,
+        permalink: setPermalink,
+        thirdNavTitle: setThirdNavTitle,
+        date: setResourceDate,
+        fileUrl: setFileUrl,
+    }
+
+    useEffect(() => {
+        if (originalCategory) setCategory(originalCategory);
+
+        const baseUrl = `${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}${originalCategory ? type === "resource" ? `/resources/${originalCategory}` : `/collections/${originalCategory}` : ''}`
+        setBaseApiUrl(baseUrl)
+
+        const fetchData = async () => {
+            // Retrieve the list of all page/resource categories for use in the dropdown options. Also sets the default category if none is specified.
+            if (type === 'resource') {
+                const resourcesResp = await axios.get(`${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/resources`);
+                const { resources } = resourcesResp.data;
+                setAllCategories(resources.map((category) => category.dirName))
+            } else if (type === 'page') {
+                const collectionsResp = await axios.get(`${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/collections`);
+                const { collections } = collectionsResp.data;
+                const collectionCategories = [''].concat(collections) // allow for selection of "Unlinked Page" category
+                setAllCategories(collectionCategories.map((category) => (
+                    {
+                      value:category,
+                      label:category ? category : 'Unlinked Page'
+                    }
+                )))
+            }
+
+            // Set component form values
+            if (isNewFile) {
+                // Set default values for new file
+                setTitle('Title')
+                setPermalink('permalink')
+                if (type === 'resource') setResourceDate(new Date().toISOString().split("T")[0])
+
+            } else {
+                // Retrieve data from an existing page/resource
+                const resp = await axios.get(`${baseUrl}/pages/${fileName}`);
+                const { content, sha: fileSha } = resp.data;
+                const base64DecodedContent = Base64.decode(content);
+                const { frontMatter, mdBody } = frontMatterParser(base64DecodedContent);
+
+                let existingLink
+                if (frontMatter.permalink) {
+                  // We want to set parts of the link by default, and only leave a portion editable by the user
+                  const { editableLink } = retrieveCollectionAndLinkFromPermalink(frontMatter.permalink)
+                  existingLink = editableLink
+                }
+
+                // File properties
+                setSha(fileSha)
+                setMdBody(mdBody)
+                setIsPost(!frontMatter.file_url)
+
+                // Front matter properties
+                setTitle(dequoteString(frontMatter.title))
+
+                setPermalink(existingLink)
+                setOriginalPermalink(existingLink)
+                setFileUrl(frontMatter.file_url)
+                setOriginalFileUrl(frontMatter.file_url)
+
+                setResourceDate(frontMatter.date)
+                setOriginalThirdNavTitle(frontMatter.third_nav_title)
+                setThirdNavTitle(frontMatter.third_nav_title)
+            }
+        }
+
+        try {
+            fetchData()
+        } catch (err) {
+            console.log(err)
+        }
+    }, [])
+
+    useEffect(() => {
+        setHasErrors(_.some(errors, (field) => field.length > 0));
+    }, [errors])
+
+    const handlePermalinkFileUrlToggle = (event) => {
+        const { target: { value } } = event;
+        if (value === 'file') {
+            setPermalink('')
+            setFileUrl(originalFileUrl ? originalFileUrl : '')
+            setIsPost(false)
+            setErrors((prevState) => ({
+                ...prevState,
+                permalink: '',
+            }))
+        } else {
+            setPermalink(originalPermalink ? originalPermalink : 'permalink')
+            setFileUrl('')
+            setIsPost(true)
+            setErrors((prevState) => ({
+                ...prevState,
+                fileUrl: '',
+            }))
+        }
+    }
+
+    const saveHandler = async () => {
+        try {
+            const fileInfo = {
+                // state
+                title,
+                permalink,
+                fileUrl,
+                date: resourceDate,
+                mdBody,
+                sha,
+                category,
+                baseApiUrl,
+                originalThirdNavTitle,
+                thirdNavTitle,
+                // props
+                originalCategory,
+                collectionPageData,
+                type,
+                fileName,
+                isNewFile,
+                siteName,
+            }
+
+            const redirectUrl = await saveFileAndRetrieveUrl(fileInfo)
+
+            // Refresh page
+            !isNewFile && window.location.reload();
+
+            setNewPageUrl(redirectUrl)
+            setRedirectToNewPage(true)
+        } catch (err) {
+            console.log(err);
+        }
+    }
+
+    const deleteHandler = async () => {
+        try {
+            const params = { sha };
+            await axios.delete(`${baseApiUrl}/pages/${fileName}`, {
+                data: params,
+            });
+
+            // Refresh page
+            window.location.reload();
+        } catch (err) {
+            console.log(err);
+        }
+    }
+
+    const changeHandler = (event) => {
+        const { id, value } = event.target;
+        const currentFileName = fileName;
+
+        const pageFileNamesExc = pageFileNames ? pageFileNames.filter((filename) => filename !== currentFileName) : null;
+
+        let errorMessage, newFileName
+        if (type === 'resource') {
+            errorMessage = validateResourceSettings(id, value);
+            newFileName = generateResourceFileName(id === "title" ? value : title, id==="date" ? value : resourceDate)
+        } else if (type === 'page') {
+            errorMessage = validatePageSettings(id, value);
+            let groupIdentifier
+            if (originalCategory) {
+                groupIdentifier = fileName.split('-')[0]
+                newFileName = generateCollectionPageFileName(value, groupIdentifier)
+            } else {
+                newFileName = generatePageFileName(value);
+            }
+        }
+    
+        if (errorMessage === '' && pageFileNamesExc && pageFileNamesExc.includes(newFileName)) {
+            setErrors((prevState) => ({
+                ...prevState,
+                title: 'This title is already in use. Please choose a different one.',
+            }));
+            idToSetterFuncMap[id](value);
+        } else {
+            setErrors((prevState) => ({
+                ...prevState,
+                [id]: errorMessage,
+            }));
+            idToSetterFuncMap[id](value);
+        }
+    }
+
+    const dropdownChangeHandler = (event) => {
+        const { id, value } = event.target;
+        idToSetterFuncMap[id](value);
+    }
+
+    const categoryDropdownHandler = (newValue) => {
+        let event;
+        if (!newValue) {
+          // Field was cleared
+          event = {
+            target: {
+              id: 'category',
+              value: '',
+            }
+          }
+        } else {
+          const { value } = newValue
+          event = {
+            target: {
+              id: 'category',
+              value,
+            }
+          }
+        }
+        
+        dropdownChangeHandler(event);
+    };
+
+    // Artificially create event from dropdown action
+    const thirdNavDropdownHandler = (newValue) => {
+        let event;
+        if (!newValue) {
+            // Field was cleared
+            event = {
+                target: {
+                    id: 'thirdNavTitle',
+                    value: '',
+                },
+            }
+        } else {
+            const { value } = newValue
+            event = {
+                target: {
+                    id: 'thirdNavTitle',
+                    value,
+                },
+            }
+        }
+
+        dropdownChangeHandler(event)
+    }
+
+    return (
+        <>
+          <div className={elementStyles.overlay}>
+            { (sha || isNewFile)
+            && (
+            <div className={elementStyles['modal-settings']}>
+              <div className={elementStyles.modalHeader}>
+                <h1>{modalTitle}</h1>
+                <button id="settings-CLOSE" type="button" onClick={settingsToggle}>
+                  <i id="settingsIcon-CLOSE" className="bx bx-x" />
+                </button>
+              </div>
+              <div className={elementStyles.modalContent}>
+                <div className={elementStyles.modalFormFields}>
+                  {/* Category */}
+                  <p className={elementStyles.formLabel}>Category Folder Name</p>
+                  <div className="d-flex text-nowrap">
+                    <CreatableSelect
+                      isClearable
+                      className="w-100"
+                      onChange={categoryDropdownHandler}
+                      isDisabled={isCategoryDisabled}
+                      value={{
+                          value: originalCategory,
+                          label: generateInitialCategoryLabel(originalCategory, isCategoryDisabled),
+                        }}
+                      options={allCategories}
+                    />
+                  </div>
+                  <span className={elementStyles.error}>{errors.category}</span>
+                  {/* Title */}
+                  <FormField
+                    title="Title"
+                    id="title"
+                    value={title}
+                    errorMessage={errors.title}
+                    isRequired={true}
+                    onFieldChange={changeHandler}
+                  />
+                  {/* Third Nav */}
+                  { 
+                    type === "page" &&
+                    originalCategory &&
+                    <>
+                        <p className={elementStyles.formLabel}>Third Nav Section</p>
+                        <div className="d-flex text-nowrap">
+                            <AsyncCreatableSelect
+                                isClearable
+                                cacheOptions
+                                defaultOptions
+                                className="w-100"
+                                onChange={thirdNavDropdownHandler}
+                                value={{
+                                    value: thirdNavTitle,
+                                    label: generateInitialThirdNavLabel(thirdNavTitle, originalCategory),
+                                }}
+                                loadOptions={loadThirdNavOptions}
+                            />
+                        </div>
+                    </>
+                  }
+                  {/* Permalink */}
+                  <FormField
+                    title="Permalink"
+                    id="permalink"
+                    value={permalink ? permalink : ''}
+                    errorMessage={errors.permalink}
+                    isRequired={true}
+                    onFieldChange={changeHandler}
+                    disabled={!isPost}
+                    fixedMessage={originalCategory ? `/${originalCategory}/${thirdNavTitle ? `${thirdNavTitle}/` : ''}` : '/'}
+                    placeholder=' '
+                  />
+                  { type === "resource" && 
+                  <ResourceFormFields 
+                    date={resourceDate}
+                    errors={errors}
+                    changeHandler={changeHandler}
+                    onToggle={handlePermalinkFileUrlToggle}
+                    isPost={isPost}
+                    setIsPost={setIsPost}
+                    siteName={siteName}
+                    fileUrl={fileUrl ? fileUrl : ''}
+                  />
+                  }
+                </div>
+                <SaveDeleteButtons 
+                  isDisabled={isNewFile ? hasErrors || (type==='resource' && category==='') : (hasErrors || !sha)}
+                  hasDeleteButton={!isNewFile}
+                  saveCallback={saveHandler}
+                  deleteCallback={() => setCanShowDeleteWarningModal(true)}
+                />
+              </div>
+            </div>
+            )}
+          </div>
+          {
+            canShowDeleteWarningModal
+            && (
+              <DeleteWarningModal
+                onCancel={() => setCanShowDeleteWarningModal(false)}
+                onDelete={deleteHandler}
+                type="resource"
+              />
+            )
+          }
+          {
+            isNewFile
+            && redirectToNewPage
+            && (
+              <Redirect
+                to={{
+                  pathname: newPageUrl
+                }}
+              />
+            )
+          }
+        </>
+    );
+}
+
+export default ComponentSettingsModal
+
 ComponentSettingsModal.propTypes = {
-  siteName: PropTypes.string.isRequired,
-  fileName: PropTypes.string.isRequired,
-  category: PropTypes.string,
-  isNewFile: PropTypes.bool.isRequired,
-  settingsToggle: PropTypes.func.isRequired,
-};
+    siteName: PropTypes.string.isRequired,
+    fileName: PropTypes.string.isRequired,
+    category: PropTypes.string,
+    isNewFile: PropTypes.bool.isRequired,
+    settingsToggle: PropTypes.func.isRequired,
+  };
